refactor: self code review
