+++ src/components/__tests__/character-class.test.ts
@@ -1,4 +1,4 @@
-import { buildPattern } from '../..';
+import { buildPattern } from '../../builders';
 import { one, oneOrMore } from '../quantifiers';
 import {
   any,
