+++ src/components/__tests__/quantifiers.test.tsx
@@ -1,4 +1,4 @@
-import { buildPattern, buildRegex } from '../..';
+import { buildPattern, buildRegex } from '../../builders';
 import { digit } from '../character-class';
 import { one, oneOrMore, optionally, zeroOrMore } from '../quantifiers';
